image_Agent/requirements.txt
cloudpickle==3.0
google-adk[eval]>=1.22.1
google-auth-oauthlib>=1.2.4
google-cloud-aiplatform[adk,agent-engines]>=1.135.0
google-cloud-logging>=3.13.0
google-cloud-storage>=3.8.0
google-genai>=1.60.0
httpx>=0.28.1
ipykernel>=7.1.0
opentelemetry-instrumentation-google-genai>=0.6b0
pydantic==2.12.5
python-dotenv>=1.2.1
requests>=2.32.5
google-cloud-secret-manager>=2.21.0
pytest>=8.0.0
